--- ThreeDeeFrontend/Pages/FilesView.razor
@@ -1,18 +1,0 @@
-@page "/"
-@using ThreeDeeFrontend.Components
-@using ThreeDeeApplication.Enums
-<MudContainer MaxWidth="MaxWidth.Large">
-    <MudText Style="padding-top: 50px;" Typo="Typo.h2">Files</MudText>
-    <div style="width: 100%; display: flex;">
-        <div style="width: 100%; margin-right: 20px; ">
-            <SearchBar
-                FilteredFilesHaveChanged="async (v) => await OnFilteredValueChanged(v)"/>
-        </div>
-        <div style="display: flex; flex-direction: row-reverse; align-items: flex-end; gap: 10px"> 
-            <FileAccessFilter FiletypeHasChanged="OnStatusButtonClicked"></FileAccessFilter>
-        </div>
-    </div>
-    
-    <FileGrid FilteredFiles="_filteredFiles"></FileGrid>
-    
-</MudContainer>
