+++ ThreeDeeFrontend/Pages/Index.razor
@@ -1,21 +1,19 @@
-@page "/index"
+@page "/"
 @using ThreeDeeFrontend.Components
+
+
 <MudContainer MaxWidth="MaxWidth.Large">
     <MudText Style="padding-top: 50px;" Typo="Typo.h2">Files</MudText>
-    <SearchBar 
-        FilteredFilesHaveChanged="async (v) => await OnFilteredValueChanged(v)"/>
-    <MudGrid Style="margin-top: 25px;">
-        @foreach (var file in _filteredFiles)
-        {
-            <MudItem xs="3" Style="padding: 0; margin: 0;">
-                <MudContainer
-                    Style="padding: 0; margin: 0;"
-                    Class="d-flex align-center justify-center mud-width-full">
-                    <FileCardPreview
-                        File="@file"
-                        ButtonClicked="OnButtonClicked"/>
-                </MudContainer>
-            </MudItem>
-        }
-    </MudGrid>
+    <div style="width: 100%; display: flex;">
+        <div style="width: 100%; margin-right: 20px; ">
+            <SearchBar
+                FilteredFilesHaveChanged="async (v) => await OnFilteredValueChanged(v)"/>
+        </div>
+        <div style="display: flex; flex-direction: row-reverse; align-items: flex-end; gap: 10px"> 
+            <FileAccessFilter FiletypeHasChanged="OnStatusButtonClicked"></FileAccessFilter>
+        </div>
+    </div>
+    
+    <FileGrid FilteredFiles="_filteredFiles"></FileGrid>
+    
 </MudContainer>
