--- ThreeDeeFrontend/Pages/SharedFile.razor
@@ -1,21 +1,0 @@
-@page "/sharedfile"
-@using ThreeDeeFrontend.Components
-<MudContainer MaxWidth="MaxWidth.Large">
-    <MudText Style="padding-top: 50px;" Typo="Typo.h2">Files</MudText>
-    <SearchBar 
-        FilteredFilesHaveChanged="async (v) => await OnFilteredValueChanged(v)"/>
-    <MudGrid Style="margin-top: 25px;">
-        @foreach (var file in _filteredSharedFiles)
-        {
-            <MudItem xs="3" Style="padding: 0; margin: 0;">
-                <MudContainer
-                    Style="padding: 0; margin: 0;"
-                    Class="d-flex align-center justify-center mud-width-full">
-                    <FileCardPreview
-                        File="@file"
-                        ButtonClicked="OnButtonClicked"/>
-                </MudContainer>
-            </MudItem>
-        }
-    </MudGrid>
-</MudContainer>
